moved all current enki source to `legacy` folder
