+++ enki/legacy/impl/any_byte_iterator.hpp
@@ -5,7 +5,7 @@
 #include <memory>
 #include <optional>
 
-#include "enki/impl/concepts.hpp"
+#include "enki/legacy/impl/concepts.hpp"
 
 namespace enki
 {
