+++ enki/legacy/impl/cstr.hpp
