+++ enki/legacy/impl/utilities.hpp
@@ -1,7 +1,7 @@
 #ifndef ENKI_UTILITIES_HPP
 #define ENKI_UTILITIES_HPP
 
-#include "enki/impl/concepts.hpp"
+#include "enki/legacy/impl/concepts.hpp"
 
 namespace enki
 {
