+++ tests/regression_tests/base_engine_array_serdes.cpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 #include <algorithm>
 
-#include "enki/base_engine.hpp"
+#include "enki/legacy/base_engine.hpp"
 
 #include <catch2/catch_test_macros.hpp>
