+++ tests/regression_tests/base_engine_composite_tests.cpp
@@ -6,7 +6,7 @@
 #include <type_traits>
 #include <unordered_map>
 
-#include "enki/base_engine.hpp"
+#include "enki/legacy/base_engine.hpp"
 
 #include <catch2/catch_test_macros.hpp>
