+++ tests/regression_tests/base_engine_range_constructible_serdes.cpp
@@ -9,7 +9,7 @@
 #include <forward_list>
 #include <list>
 
-#include "enki/base_engine.hpp"
+#include "enki/legacy/base_engine.hpp"
 
 #include <catch2/catch_test_macros.hpp>
