+++ tests/regression_tests/base_engine_requirements.cpp
@@ -2,7 +2,7 @@
 #include <concepts>
 #include <type_traits>
 
-#include "enki/base_engine.hpp"
+#include "enki/legacy/base_engine.hpp"
 
 #include "catch2/catch_test_macros.hpp"
