+++ tests/regression_tests/manager_requirements.cpp
@@ -7,7 +7,7 @@
 #include <variant>
 #include <vector>
 
-#include "enki/manager.hpp"
+#include "enki/legacy/manager.hpp"
 
 #include "catch2/catch_test_macros.hpp"
